Refatorando handler de erro: centralizando na criacao do app para evitar duplicacao de codigo
