+++ app/resources/campus.py
@@ -1,12 +1,10 @@
 from flask import make_response, Blueprint, request, jsonify
 from ..controllers.campus import CampusController
 from ..errors import ErroInterno, UsoInvalido, TipoErro
-from . import generic_handler, login_required
+from . import login_required
 
 
 bp = Blueprint("campus", __name__, url_prefix="/campus")
-bp.register_error_handler(ErroInterno, generic_handler)
-bp.register_error_handler(UsoInvalido, generic_handler)
 
 
 @bp.route("/", methods=["POST"])
