+++ app/resources/docs.py
@@ -1,11 +1,8 @@
 from flask import Blueprint, render_template
-from ..errors import ErroInterno, UsoInvalido, TipoErro
-from . import generic_handler
+from ..errors import ErroInterno, TipoErro
 
 
 bp = Blueprint("docs", __name__, url_prefix="/apidocs")
-bp.register_error_handler(ErroInterno, generic_handler)
-bp.register_error_handler(UsoInvalido, generic_handler)
 
 
 @bp.route("/", methods=("GET",))
